# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Clientes",
        "description": "API para la gestión de clientes",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener lista de clientes",
                "description": "Retorna lista de clientes paginada",
                "operationId": "clienteIndex",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear nuevo cliente",
                "description": "Almacena un nuevo cliente y retorna los datos",
                "operationId": "clienteStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_nombre",
                                    "cliente_razonsocial"
                                ],
                                "properties": {
                                    "cliente_nombre": {
                                        "type": "string",
                                        "example": "Empresa XYZ"
                                    },
                                    "cliente_razonsocial": {
                                        "type": "string",
                                        "example": "Empresa XYZ S.A."
                                    },
                                    "limite_credito": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    },
                                    "credito_habilitado": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "credito_utilizado": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Mostrar información de un cliente",
                "description": "Retorna los datos de un cliente específico",
                "operationId": "clienteShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar cliente existente",
                "description": "Actualiza los datos de un cliente específico y retorna los datos actualizados",
                "operationId": "clienteUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cliente_nombre": {
                                        "type": "string",
                                        "example": "Empresa XYZ Actualizada"
                                    },
                                    "cliente_razonsocial": {
                                        "type": "string",
                                        "example": "Empresa XYZ S.A. Actualizada"
                                    },
                                    "limite_credito": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000
                                    },
                                    "credito_habilitado": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "credito_utilizado": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar cliente",
                "description": "Elimina un cliente específico",
                "operationId": "clienteDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cliente eliminado exitosamente"
                    },
                    "404": {
                        "description": "Cliente no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes/search": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Buscar clientes",
                "description": "Busca clientes por nombre",
                "operationId": "clienteSearch",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre a buscar",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Cliente": {
                "required": [
                    "cliente_nombre",
                    "cliente_razonsocial"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "cliente_nombre": {
                        "type": "string",
                        "example": "Empresa XYZ"
                    },
                    "cliente_razonsocial": {
                        "type": "string",
                        "example": "Empresa XYZ S.A."
                    },
                    "limite_credito": {
                        "type": "number",
                        "format": "float",
                        "example": 10000
                    },
                    "credito_habilitado": {
                        "type": "boolean",
                        "example": true
                    },
                    "credito_utilizado": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Clientes",
            "description": "Clientes"
        }
    ]
}